Move solidarity.js to solidarity.ts
