+++ tests/commands/check.js
@@ -1,5 +1,5 @@
 import test from 'ava'
-import checkCommand from '../../src/commands/solidarity'
+import checkCommand from '../../dist/commands/solidarity'
 
 test('Snapshot check default command', t => {
   t.snapshot(checkCommand)
