+++ tests/commands/snapshots/check.js.md
@@ -10,7 +10,5 @@ Generated by [AVA](https://ava.li).
 
     {
       description: 'Check environment rules',
-      run: AsyncFunction [
-        undefined,
-      ],
+      run: Function {},
     }
